Update best-practices.md
This part is redundant. We already recommend not scoping policies to all users, all apps, and block access. No need to call out "same policy but with all other conditions." By the way, we actually block admins from creating this policy unless there is an exclusion (either in users or conditions).