Adding note for specific use case warning
I recommend this Note be added as this scenario actually occurred with one of my customers. They were staging an access package for guests for a highly sensitive project, but used a Microsoft 365 group instead of a security group. All the participants were notified by email and they tried to access the resources before they were ready.